Corrected mixup between integration and unit tests
